Custom field with a default value
This is a work in progress. I feel the complexity 
of this implementation is already quite high and 
likely to grow as the initial list of options is added
i.e. validation, field type, etc..

As such it may be worth thinking about alternatives,
even if that is simply extracting the code to a gem
and going nuts with the testing there.