Clean up CustomFields code and add unit tests
It is expected that a change to the fields config
will result in a new portfolio (version) being created
which will then create a new `Custom` class definition.

Load testing is needed to understand the impact of
this runtime creation of class objects as they are
being stored on the Trade object as constants.

The alternative would be to create and then discard
the class objects overtime they are needed, this 
would avoid runtime creation of constants at the
cost of performance.