Start work on user permissions using `pundit` gem
Implementation of CREATE_PORTFOLIO permissions.

In addition a `SUPER_ADMIN` role has been created
which would give blanket access to all functionality.

It is expected that roles will be created for each
business with default permissions, and that a super admin
or paying custom would be able to adjust the permissions
on a given role or even create new sub roles based on
the available permissions.