Start work on building demo trades and positions
This is designed to give the user feel of what 
can be done.

NOTE - This expects a single demo portfolio to be 
shared amongst all users (once sharing is enabled)
but could easily be created for separate portfolios
with separate trade positions if required.

This expects the portfolio to already exist before the 
script is run.