Build example trades and portfolio display
Add rake task to generate example trades based off 
ticker to valuation map. This includes tracking the 
security and creating a set of historical trades
which make up the life cycle of the position.

The next step is the ability to share the portfolio
with other users/businesses so that they (Elly) can 
see the progress.

An alternative would be to add this portfolio for 
Elly and then deal with sharing later?